Ticks and server
Significant changes to workflows to ensure they work with the new client server connection.

Went through and replaced as many descriptions of "angles" with "ticks" as I could find and were appropriate. Sometimes angles is fine when describing "a PPM angle", but at other times it requires stating ticks to differentiate the actual rotation angle the motor sees from the double angle (tick) that the PPM is labeled with.